Don't double-delete caches by element type

+++ src/services/TemplateCacheService.php
@@ -287,6 +287,11 @@ class TemplateCacheService extends BaseApplicationComponent
 	 */
 	public function deleteCachesByElementType($elementType)
 	{
+		if (!empty($this->_deletedCachesByElementType[$elementType]))
+		{
+			return false;
+		}
+
 		$affectedRows = craft()->db->createCommand()->delete(static::$_templateCachesTable, array('type = :type'), array(':type' => $elementType));
 
 		$this->_deletedCachesByElementType[$elementType] = true;
